requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.1
django-widget-tweaks==1.4.5
idna==2.8
newsapi-python==0.2.6
pkg-resources==0.0.0
python-dateutil==2.8.1
python-http-client==3.2.1
pytz==2019.1
requests==2.22.0
sendgrid==6.1.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
